Added parameters csv file that contains falciparum and vivax specific parameter sets.
These are now selected in the get_parameters() function by specificying parasite == "falciparum" or "vivax". The functions then selects the parasite-specific values and combines then withe the non-parasite specific values.

In order to allow the vivax code to run without full functionality, we have included additional falciparum parameters in the vivax parameter set. These will be removed in the following stages.

Added tests to check parameters are behaving as expected.