antiparasite immunity function shapes the duration of subpatent infection/rate of recovery from subpatent infection. These values are now assigned during the variable creation and calculation of recovery rates.
In previous iterations I have kept ID in the vivax model, where it acts as antiparasite immunity. For clarity of thought, I've now added IAA and IAM antiparasite immunities, including their decay, boosting, maternal immunity and reset at death.

We must also adjust the prevalence rendering, which for p.f is based on detectability immunity, and for p.v is explicit in the human states. p_detect_lm is also inappropriate for p.v as the lm_detection is explicit.

du, and ID parameters removed for vivax.

Tests are added for du rates and for prevalence rendering.