I've renamed patent infections as light microscopy detectable infections or lm_det, to avoid the confusion of what exactly patent means.
I've also removed the patent/lm_det infections <- NULL line from the falciparum code, but in doing so, have replicated some infection functions for falciparum and vivax so that schedule_infections may (for vivax) or may not (for falciparum) use the lm_det subset. I've also reordered the arguments for that function so that the lm_det_infections can be at the end, but make sense in the context of other arguments.